chore: 코드 컴포넌트 분리 및 UI 구분
