--- apps/client/src/app/search/_components/FilterBar/index.ts
@@ -1,1 +1,0 @@
-export * from "./FilterBar";
